requirements.txt
antiorm==1.2.1
Click==7.0
db==0.1.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
six==1.14.0
SQLAlchemy==1.3.12
Werkzeug==0.16.0
